requirements/base.txt
Django==5.0.4
python-decouple==3.8
psycopg2==2.9.9
djangorestframework==3.15.1
django-mptt==0.16.0
django-ckeditor==6.7.1
celery==5.4.0
redis==5.0.4
django-filter==24.2